+++ plugins/module_utils/api.py
@@ -29,13 +29,15 @@ __metaclass__ = type
 #
 
 
+import json
 import os
 from functools import partial
 from ansible.module_utils._text import to_native
 from ansible.module_utils.six import iteritems
 from ansible.module_utils._text import to_text
 from ansible.module_utils.basic import env_fallback
-from ansible.module_utils.common.validation import check_type_dict
+from ansible.module_utils.common.validation import check_type_dict, safe_eval
+from ansible.module_utils.six import string_types
 
 try:
     from infoblox_client.connector import Connector
@@ -313,6 +315,20 @@ class WapiModule(WapiBase):
         if (ib_obj_type == NIOS_MEMBER):
             proposed_object = member_normalize(proposed_object)
 
+        # checks if the 'text' field has to be updated for the TXT Record
+        if (ib_obj_type == NIOS_TXT_RECORD):
+            text_obj = proposed_object["text"]
+            if text_obj.startswith("{"):
+                try:
+                    text_obj = json.loads(text_obj)
+                    txt = text_obj['new_text']
+                except Exception:
+                    (result, exc) = safe_eval(text_obj, dict(), include_exceptions=True)
+                    if exc is not None:
+                        raise TypeError('unable to evaluate string as dictionary')
+                    txt = result['new_text']
+                proposed_object['text'] = txt
+
         # checks if the name's field has been updated
         if update and new_name:
             proposed_object['name'] = new_name
@@ -510,12 +526,13 @@ class WapiModule(WapiBase):
 
         update = False
         old_name = new_name = None
+        old_ipv4addr_exists = old_text_exists = False
         if ('name' in obj_filter):
             # gets and returns the current object based on name/old_name passed
             try:
                 name_obj = check_type_dict(obj_filter['name'])
-                old_name = name_obj['old_name']
-                new_name = name_obj['new_name']
+                old_name = name_obj['old_name'].lower()
+                new_name = name_obj['new_name'].lower()
             except TypeError:
                 name = obj_filter['name']
 
@@ -529,11 +546,12 @@ class WapiModule(WapiBase):
                 if ib_obj:
                     obj_filter['name'] = new_name
                 else:
-                    test_obj_filter['name'] = new_name
-                    ib_obj = self.get_object(ib_obj_type, test_obj_filter, return_fields=list(ib_spec.keys()))
+                    raise Exception("object with name: '%s' is not found" % (old_name))
                 update = True
                 return ib_obj, update, new_name
             if (ib_obj_type == NIOS_HOST_RECORD):
+                # to fix the sanity issue
+                name = obj_filter['name']
                 # to check only by name if dns bypassing is set
                 if not obj_filter['configure_for_dns']:
                     test_obj_filter = dict([('name', name)])
@@ -553,6 +571,7 @@ class WapiModule(WapiBase):
                 try:
                     ipaddr_obj = check_type_dict(obj_filter['ipv4addr'])
                     ipaddr = ipaddr_obj.get('old_ipv4addr')
+                    old_ipv4addr_exists = True
                 except TypeError:
                     ipaddr = obj_filter['ipv4addr']
                 test_obj_filter['ipv4addr'] = ipaddr
@@ -560,8 +579,20 @@ class WapiModule(WapiBase):
                 # resolves issue where multiple txt_records with same name and different text
                 test_obj_filter = obj_filter
                 try:
-                    text_obj = check_type_dict(obj_filter['text'])
-                    txt = text_obj['old_text']
+                    text_obj = obj_filter['text']
+                    if text_obj.startswith("{"):
+                        try:
+                            text_obj = json.loads(text_obj)
+                            txt = text_obj['old_text']
+                            old_text_exists = True
+                        except Exception:
+                            (result, exc) = safe_eval(text_obj, dict(), include_exceptions=True)
+                            if exc is not None:
+                                raise TypeError('unable to evaluate string as dictionary')
+                            txt = result['old_text']
+                            old_text_exists = True
+                    else:
+                        txt = text_obj
                 except TypeError:
                     txt = obj_filter['text']
                 test_obj_filter['text'] = txt
@@ -569,26 +600,55 @@ class WapiModule(WapiBase):
             else:
                 test_obj_filter = obj_filter
             ib_obj = self.get_object(ib_obj_type, test_obj_filter.copy(), return_fields=list(ib_spec.keys()))
+
+            # prevents creation of a new A record with 'new_ipv4addr' when A record with a particular 'old_ipv4addr' is not found
+            if old_ipv4addr_exists and ib_obj is None:
+                raise Exception("A Record with ipv4addr: '%s' is not found" % (ipaddr))
+            # prevents creation of a new TXT record with 'new_text' when TXT record with a particular 'old_text' is not found
+            if old_text_exists and ib_obj is None:
+                raise Exception("TXT Record with text: '%s' is not found" % (txt))
         elif (ib_obj_type == NIOS_A_RECORD):
             # resolves issue where multiple a_records with same name and different IP address
             test_obj_filter = obj_filter
             try:
                 ipaddr_obj = check_type_dict(obj_filter['ipv4addr'])
                 ipaddr = ipaddr_obj.get('old_ipv4addr')
+                old_ipv4addr_exists = True
             except TypeError:
                 ipaddr = obj_filter['ipv4addr']
             test_obj_filter['ipv4addr'] = ipaddr
+            # prevents creation of a new A record with 'new_ipv4addr' when A record with a particular 'old_ipv4addr' is not found
+            if old_ipv4addr_exists and ib_obj is None:
+                raise Exception("A Record with ipv4addr: '%s' is not found" % (ipaddr))
             ib_obj = self.get_object(ib_obj_type, test_obj_filter.copy(), return_fields=list(ib_spec.keys()))
+            # prevents creation of a new A record with 'new_ipv4addr' when A record with a particular 'old_ipv4addr' is not found
+            if old_ipv4addr_exists and ib_obj is None:
+                raise Exception("A Record with ipv4addr: '%s' is not found" % (ipaddr))
         elif (ib_obj_type == NIOS_TXT_RECORD):
             # resolves issue where multiple txt_records with same name and different text
             test_obj_filter = obj_filter
             try:
-                text_obj = check_type_dict(obj_filter['text'])
-                txt = text_obj['old_text']
+                text_obj = obj_filter(['text'])
+                if text_obj.startswith("{"):
+                    try:
+                        text_obj = json.loads(text_obj)
+                        txt = text_obj['old_text']
+                        old_text_exists = True
+                    except Exception:
+                        (result, exc) = safe_eval(text_obj, dict(), include_exceptions=True)
+                        if exc is not None:
+                            raise TypeError('unable to evaluate string as dictionary')
+                        txt = result['old_text']
+                        old_text_exists = True
+                else:
+                    txt = text_obj
             except TypeError:
                 txt = obj_filter['text']
             test_obj_filter['text'] = txt
             ib_obj = self.get_object(ib_obj_type, test_obj_filter.copy(), return_fields=list(ib_spec.keys()))
+            # prevents creation of a new TXT record with 'new_text' when TXT record with a particular 'old_text' is not found
+            if old_text_exists and ib_obj is None:
+                raise Exception("TXT Record with text: '%s' is not found" % (txt))
         elif (ib_obj_type == NIOS_ZONE):
             # del key 'restart_if_needed' as nios_zone get_object fails with the key present
             temp = ib_spec['restart_if_needed']